+++ lib/plutus/transformer.rb
@@ -12,7 +12,7 @@ module Plutus
           first_name:    row[0],
           last_name:     row[1],
           annual_salary: row[2].to_i,
-          super_rate:    row[3].to_f,
+          super_rate:    row[3].to_d,
           month_period:  row[4]
         }
       end
